api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
pydantic==2.*
SQLAlchemy==2.*
alembic==1.*
psycopg2-binary==2.*
httpx==0.27.*
tenacity==9.*
tencentcloud-sdk-python==3.*
